% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app_ui}
\alias{app_ui}
\alias{app_server}
\title{Creates the User Interface for the "Old Faithful Geyser Data" Application}
\usage{
app_ui()

app_server(input, output)
}
\arguments{
\item{input}{The input variable from the user interface.}

\item{output}{The output variable from the user interface.}
}
\description{
This function creates the user interface for the "Old Faithful Geyser Data" application.
It generates a fluid page that contains a title panel, a sidebar panel with a slider
input for the number of bins, and a main panel that displays a plot of the
generated distribution.
}
\examples{
\dontrun{
  shinyApp(ui = app_ui, server = app_server)
}
}
