% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_histogram.R
\name{create_histogram}
\alias{create_histogram}
\title{Create a histogram of the input data}
\usage{
create_histogram(x, num_bins)
}
\arguments{
\item{x}{a numeric vector of data values}

\item{num_bins}{an integer specifying the number of bins to use in the histogram}
}
\value{
a histogram plot of the input data
}
\description{
Create a histogram of the input data
}
\examples{
create_histogram(rnorm(100), 10)

}
